Real Fixed Chat Callback issue but that still needs more investigating..
The fix was kinda odd, I had to move where RegisterRTTChatCallback was being called to where we enable RTT and RegisterRTTPresenceCallback. What I dont get is why this has to be in the same execute line in order for the callbacks to work? They both just add callbacks to a TMap that gets checked when the websocket gets a OnMessage callback. Seems suspicious...