Chat callback clean up + Bug fix with Presence erasing the username
The bug would occur when the user is logged in and a friend is offline, when that friend goes online then the name gets erased but shows "Online". The reason this happened was because of Unreal's janky json reader plugin, for some reason it wont read in the "name" field no matter what object you throw at it (if its the whole data or the object that holds the name field. Since this is a callback for presence I just bypass this check and feed in the appropriate field.