Took out verify lobby ID
Run script that I've set up wasn't working. for some reason when I send a string as {lobbyID:<lobbyIDHere>} it would return null. Code is kinda hard to determine when I can breakpoint so I yeeted it out and just added a failure callback for Joining Lobby if it doesnt work.